refactor tenant flow and keep the public classes intact as prev implementation
