+++ components/validation/src/main/java/org/wso2/carbon/identity/x509Certificate/validation/RevocationStatus.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.x509Certificate.validation.model;
+package org.wso2.carbon.identity.x509Certificate.validation;
 
 /**
  * This is used to get Revocation Status message.
