+++ components/validation/src/main/java/org/wso2/carbon/identity/x509Certificate/validation/X509CertificateValidationConstants.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.identity.x509Certificate.validation.constant;
+package org.wso2.carbon.identity.x509Certificate.validation;
 
 /**
  * This holds the X509 Certificate validation constants.
